[BOOK-181] refactor: apis - 애플 및 카카오 로그인 전략 리팩토링
