+++ apis/src/main/kotlin/org/yapp/apis/auth/dto/request/SaveAppleRefreshTokenRequest.kt
@@ -4,7 +4,7 @@ import io.swagger.v3.oas.annotations.media.Schema
 import jakarta.validation.constraints.NotBlank
 import jakarta.validation.constraints.NotNull
 import org.yapp.apis.user.dto.response.CreateUserResponse
-import org.yapp.apis.auth.strategy.AppleAuthCredentials
+import org.yapp.apis.auth.strategy.signin.AppleAuthCredentials
 import java.util.UUID
 
 @Schema(
