+++ apis/src/main/kotlin/org/yapp/apis/auth/strategy/signin/SignInCredentials.kt
@@ -1,20 +1,20 @@
-package org.yapp.apis.auth.strategy
+package org.yapp.apis.auth.strategy.signin
 
 import org.yapp.domain.user.ProviderType
 
-sealed class AuthCredentials {
+sealed class SignInCredentials {
     abstract fun getProviderType(): ProviderType
 }
 
 data class KakaoAuthCredentials(
     val accessToken: String
-) : AuthCredentials() {
+) : SignInCredentials() {
     override fun getProviderType(): ProviderType = ProviderType.KAKAO
 }
 
 data class AppleAuthCredentials(
     val idToken: String,
     val authorizationCode: String
-) : AuthCredentials() {
+) : SignInCredentials() {
     override fun getProviderType(): ProviderType = ProviderType.APPLE
 }
