+++ apis/src/main/kotlin/org/yapp/apis/auth/strategy/signin/SignInStrategyResolver.kt
@@ -1,14 +1,14 @@
-package org.yapp.apis.auth.strategy
+package org.yapp.apis.auth.strategy.signin
 
 import org.springframework.stereotype.Service
 import org.yapp.apis.auth.exception.AuthErrorCode
 import org.yapp.apis.auth.exception.AuthException
 
 @Service
-class AuthStrategyResolver(
-    private val strategies: List<AuthStrategy>
+class SignInStrategyResolver(
+    private val strategies: List<SignInStrategy>
 ) {
-    fun resolve(credentials: AuthCredentials): AuthStrategy {
+    fun resolve(credentials: SignInCredentials): SignInStrategy {
         return strategies.find { it.getProviderType() == credentials.getProviderType() }
             ?: throw AuthException(
                 AuthErrorCode.UNSUPPORTED_PROVIDER_TYPE,
